Voice-To-Voice-Chatbot/requirements.txt
openai-whisper 
groq 
gtts 
gradio
